stylelint

+++ stylelint.config.js
@@ -113,7 +113,7 @@ module.exports = {
         'selector-pseudo-class-no-unknown': true,
         'selector-pseudo-class-parentheses-space-inside': 'never',
         'selector-pseudo-element-case': 'lower',
-        'selector-pseudo-element-colon-notation': 'single',
+        'selector-pseudo-element-colon-notation': 'double',
         'selector-pseudo-element-no-unknown': true,
         'selector-max-type': [0, { ignoreTypes: ['use', 'a', 'svg'] }],
         'selector-no-vendor-prefix': true,
